repo/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg[binary]
alembic
pydantic
pydantic-settings
argon2-cffi
python-multipart
cryptography
httpx
pytest
pytest-asyncio
pytest-cov
anyio
greenlet
Pillow
